pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pocket-jwt-validate"
version = "0.0.1"
authors = [
  { name="Alex Dobrushskiy", email="adobrush@asu.edu" },
]
description = "JWT validation for Pocket projects"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'pyjwt >= 2.4.0',
  'cryptography >= 37.0.0'
]
